+++ src/components/Input/index.js
@@ -78,12 +78,17 @@ class Input extends React.Component {
     );
   }
 
+  renderNote({note, testSection}) {
+    return (<div className='oui-form-note' data-test-section={ testSection && testSection + '-note' }>{note}</div>);
+  }
+
   render() {
     const {
       label,
       displayError,
       testSection,
       isOptional,
+      note,
     } = this.props;
 
     if (label) {
@@ -98,11 +103,20 @@ class Input extends React.Component {
             </div>
             { this.renderInput(this.props) }
           </Label>
+          { note && this.renderNote(this.props)}
         </div>
       );
     }
 
+    if (note) {
+      return [
+        this.renderInput(this.props),
+        this.renderNote(this.props),
+      ];
+    }
+
     return this.renderInput(this.props);
+
   }
 }
 
@@ -139,6 +153,8 @@ Input.propTypes = {
    * Min value for the `input`. Should be used only when `type` is `number`.
    */
   min: PropTypes.number,
+  /** Form note for the input */
+  note: PropTypes.string,
   /**
    * Function that fires when the input loses focus. It fires regardless of
    * whether the value has changed.
@@ -171,6 +187,10 @@ Input.propTypes = {
   ]).isRequired,
   /** Text within the input */
   value: PropTypes.string,
+};
+
+Input.defaultProps = {
+  note: null,
 };
 
 export default Input;