+++ src/components/Textarea/Textarea.story.js
@@ -17,16 +17,36 @@ stories
   ));
 
 stories
-  .add('textarea with knobs', withInfo()(() => (<div>
-    <Textarea
-      isDisabled={ boolean('isDisabled', false) }
-      defaultValue='Delete this default value and see the placeholder'
-      placeholder={ text('placeHolder', 'Enter a comment') }
-      onBlur={ action('Textarea: onBlur') }
-      onChange={ action('Textarea: onChange') }
-      onFocus={ action('Textarea: onFocus') }
-      onInput={ action('Textarea: onInput') }
-      onKeyDown={ action('Textarea: onKeyDown') }
-    />
-  </div>)));
+  .add('textarea with knobs', withInfo()(() => {
+    return (
+      <div>
+        <Textarea
+          isDisabled={ boolean('isDisabled', false) }
+          defaultValue='Delete this default value and see the placeholder'
+          placeholder={ text('placeHolder', 'Enter a comment') }
+          onBlur={ action('Textarea: onBlur') }
+          onChange={ action('Textarea: onChange') }
+          onFocus={ action('Textarea: onFocus') }
+          onInput={ action('Textarea: onInput') }
+          onKeyDown={ action('Textarea: onKeyDown') }
+        />
+      </div>);
+  })).add('With label and note', () => {
+    return (
+      <Textarea
+        label="Field label"
+        note="A short description or note about this field."
+        placeholder="Just a placeholder"
+        type="text"
+      />);
+  }).add('Error state', () => {
+    return (
+      <Textarea
+        label="Field label"
+        displayError={ true }
+        note="A short description or note about this field."
+        placeholder="Just a placeholder"
+        type="text"
+      />);
+  });